Split out another loader
