+++ packages/blade/private/shell/loaders.ts
@@ -1,4 +1,4 @@
-import { readFile } from 'node:fs/promises';
+import { cp, readFile } from 'node:fs/promises';
 import path from 'node:path';
 import { compile } from '@mdx-js/mdx';
 import withToc from '@stefanprobst/rehype-extract-toc';
@@ -8,6 +8,7 @@ import type * as esbuild from 'esbuild';
 import YAML from 'js-yaml';
 import MagicString from 'magic-string';
 
+import { outputDirectory, publicDirectory } from '@/private/shell/constants';
 import {
   type ExportItem,
   type TotalFileList,
@@ -17,6 +18,12 @@ import {
   getFileList,
   wrapClientExport,
 } from '@/private/shell/utils';
+import {
+  transformToCloudflareOutput,
+  transformToNetlifyOutput,
+  transformToVercelBuildOutput,
+} from '@/private/shell/utils/providers';
+import type { DeploymentProvider } from '@/private/universal/types/util';
 import { generateUniqueId } from '@/private/universal/utils/crypto';
 
 export const getClientReferenceLoader = (): esbuild.Plugin => ({
@@ -224,6 +231,38 @@ export const getMdxLoader = (
         contents: tsx,
         loader: 'tsx',
       };
+    });
+  },
+});
+
+export const getProviderLoader = (
+  environment: 'development' | 'production',
+  provider: DeploymentProvider,
+): esbuild.Plugin => ({
+  name: 'Provider Loader',
+  setup(build) {
+    build.onEnd(async (result) => {
+      if (result.errors.length > 0 || environment !== 'production') return;
+
+      // Copy hard-coded static assets into output directory.
+      if (await exists(publicDirectory)) {
+        await cp(publicDirectory, outputDirectory, { recursive: true });
+      }
+
+      switch (provider) {
+        case 'cloudflare': {
+          await transformToCloudflareOutput();
+          break;
+        }
+        case 'netlify': {
+          await transformToNetlifyOutput();
+          break;
+        }
+        case 'vercel': {
+          await transformToVercelBuildOutput();
+          break;
+        }
+      }
     });
   },
 });
